Typed casted the int to double- to get rid of error for not able to get higher power than 34

+++ BinExpanV001.c
@@ -2,23 +2,23 @@
 #include <stdio.h>
 #include<conio.h>
 
-int factorial(int pwrValue);
+double factorial(int pwrValue);
 
 int main()
 {
-    int pwr, newValue, numerator, denominator;
+    double pwr, newValue, numerator, denominator;
     printf("\nEnter the power of (a+b) : ");
-    scanf("%d", &pwr);
+    scanf("%lf", &pwr);
 
-    for (int i = 0; i <= pwr; i++)
+    for (double i = 0; i <= pwr; i++)
     {
-        int condition = pwr -i;
+        double condition = pwr -i;
         numerator = factorial(pwr);
         denominator = factorial(pwr - i) * factorial(i);
         newValue = numerator / denominator;
-        printf(" %d",newValue);
-        printf("a^%d",pwr -i);
-        printf("b^%d",i);
+        printf(" %.2lf",newValue);
+        printf("a^%.2lf",pwr -i);
+        printf("b^%.2lf",i);
         (condition==0 && i==pwr)?printf("\n***End of the Expansion***"):printf(" + "); //Condition applied to remove the plus sign at the end
         //printf(" %d(a^%d b^%d)  +  ", newValue, pwr - i, i);              //verson 001 update
     }
@@ -26,7 +26,7 @@ int main()
     return 0;
 }
 
-int factorial(int pwrValue)
+double factorial(int pwrValue)
 {
     if (pwrValue == 0 || pwrValue < 0)
     {
